refactor: :hammer: move company creation form logic to separate component
