+++ src/components/company/index.ts
@@ -1,0 +1,1 @@
+export { CompanyCreationForm } from './CompanyCreationForm';
